+++ packages/lodim/example/example.dart
@@ -1,5 +1,3 @@
-// ignore_for_file: avoid_print, avoid_redundant_argument_values
-
 import 'package:lodim/lodim.dart';
 
 void main() {